requirements.txt
pyyaml>=6.0
regex>=2023.0.0
tabulate>=0.9.0
colorama>=0.4.6
pytest>=7.0.0